Move init code of sdl2/opengl from main crate to gb-lcd.
